Convert BrowserDetector to TS
