--- src/BrowserDetector/index.js
@@ -1,2 +1,0 @@
-import BrowserDetector from './BrowserDetector';
-export default BrowserDetector;
