+++ src/BrowserDetector/index.ts
